# openapi.yaml
openapi: 3.0.3
info:
  title: Suno API
  description: |
    A REST API wrapper around Suno.ai that enables programmatic song generation, audio extension,
    stem extraction, lyrics generation, and more. Built with Next.js and TypeScript.

    ## Authentication
    All generation endpoints require a valid Suno session cookie configured via the `SUNO_COOKIE`
    environment variable on the server. No per-request auth header is needed by API consumers.

    ## Credits
    - Song generation (generate, custom_generate, generate_sounds, extend_audio): **10 credits**
    - Concatenation: **5 credits**
    - Read-only endpoints (get, clip, models, get_limit, etc.): **free**
    - Lyrics generation: **free**

    ## Status Flow
    ```
    submitted → queued → streaming → complete
                       ↘ error
    ```
    Poll `/api/get?ids=<id>` until `status` is `complete` or `error`.
  version: "1.0.0"
  contact:
    name: Suno API
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/generate_similar:
    post:
      tags: [Generation]
      summary: Generate a song inspired by a reference audio clip
      description: |
        Provide an existing Suno clip UUID as a style reference. The API generates
        2 new songs that follow the same style, key, and feel as the reference — starting
        from scratch rather than extending it.

        Optionally override lyrics (`prompt`), style tags, and title. If omitted, Suno
        derives style automatically from the reference clip.

        Internally this uses `continue_clip_id` with `continue_at=0`, which tells Suno
        to anchor the new generation to the reference clip's style from timestamp 0.

        **Costs 10 credits.**
      operationId: generateSimilar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSimilarRequest'
            examples:
              styleOnly:
                summary: Pure style reference — no overrides
                value:
                  reference_audio_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              withOverrides:
                summary: Style reference with custom lyrics and tags
                value:
                  reference_audio_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  prompt: "[Verse]\nNew words in the same vibe\n[Chorus]\nKeep that feeling alive"
                  tags: "synthwave, melancholic, 80s"
                  title: "Echoes of the Original"
                  wait_audio: true
      responses:
        '200':
          description: Array of 2 newly generated clips inspired by the reference
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AudioInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/analyze_audio:
    post:
      tags: [Generation]
      summary: Analyze your own audio file and extract style info
      description: |
        Upload your own audio file (or provide a URL) to extract BPM, musical key,
        genre tags, and duration from its embedded metadata (ID3/Vorbis/MP4 tags).

        Returns `suggested_tags` and `suggested_prompt` that are ready to pass
        directly to `POST /api/custom_generate` or `POST /api/generate_from_audio`.

        **Accepts either:**
        - `multipart/form-data` with a `file` field (max 50 MB)
        - `application/json` with a `url` field pointing to a remote audio file

        **Note:** BPM and key are only extracted when they are embedded in the file's
        tags (common for files exported from DAWs or tagged in iTunes/MusicBrainz).
        Files without embedded tags will return `null` for those fields.
      operationId: analyzeAudio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Audio file to analyze (mp3, wav, flac, ogg, m4a, aac …)
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  description: Public URL of the audio file to analyze.
                  example: "https://example.com/my-song.mp3"
      responses:
        '200':
          description: Extracted audio analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioAnalysis'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: File too large (over 50 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/generate_from_audio:
    post:
      tags: [Generation]
      summary: Analyze your audio then generate a similar song in Suno
      description: |
        One-shot endpoint: provide your own audio file or URL, and the API will:
        1. Extract BPM, key, genre, and other style attributes from its metadata
        2. Build Suno-compatible tags automatically
        3. Call `custom_generate` with those tags (plus any overrides you provide)
        4. Return both the `analysis` and the generated `clips`

        **Accepts either:**
        - `multipart/form-data` with a `file` field + optional text fields
        - `application/json` with a `url` field + optional override fields

        **Costs 10 credits.**
      operationId: generateFromAudio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GenerateFromAudioForm'
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateFromAudioJson'
            examples:
              urlOnly:
                summary: URL only — fully automatic
                value:
                  url: "https://example.com/my-song.mp3"
              urlWithOverrides:
                summary: URL with custom lyrics and extra tags
                value:
                  url: "https://example.com/my-song.mp3"
                  title: "My New Track"
                  prompt: "[Verse]\nNew lyrics here\n[Chorus]\nFeel the vibe"
                  extra_tags: "dark, melancholic"
                  wait_audio: true
      responses:
        '200':
          description: Analysis result and generated clips
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateFromAudioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '413':
          description: File too large (over 50 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─────────────────────────────────────────────
  # OPENAI-COMPATIBLE
  # ─────────────────────────────────────────────
...
components:
  schemas:
...
    GenerateSimilarRequest:
      type: object
      required: [reference_audio_id]
      properties:
        reference_audio_id:
          type: string
          format: uuid
          description: UUID of the existing Suno clip to use as a style reference.
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        prompt:
          type: string
          description: |
            New lyrics for the generated song. Omit to let Suno derive lyrics
            automatically from the reference clip's style.
          example: "[Verse]\nNew words in the same vibe\n[Chorus]\nKeep that feeling alive"
        tags:
          type: string
          description: |
            Style/genre tags. Omit to inherit the reference clip's style automatically.
          example: "synthwave, melancholic, 80s"
        title:
          type: string
          description: Title for the new song. Defaults to empty if not provided.
          example: "Echoes of the Original"
        make_instrumental:
          type: boolean
          description: Generate without vocals.
          default: false
        model:
          type: string
          description: Suno model ID. Defaults to the auto-detected latest model.
          example: "chirp-crow"
        wait_audio:
          type: boolean
          description: Block until audio URLs are ready.
          default: false
        negative_tags:
          type: string
          description: Comma-separated styles to exclude.
          example: "heavy metal, distortion"
        vocal_gender:
          type: string
          enum: [male, female]
          description: Preferred vocal gender.
        weirdness:
          type: integer
          minimum: 0
          maximum: 100
          description: Creativity level. 0 = conventional, 100 = avant-garde, 50 = balanced.
          default: 50
        style_influence:
          type: integer
          minimum: 0
          maximum: 100
          description: How strictly tags influence the output. 50 = balanced.
          default: 50
        persona_id:
          type: string
          format: uuid
          description: Voice-cloning persona UUID.
    AudioAnalysis:
      type: object
      description: Style attributes extracted from an audio file's embedded metadata.
      properties:
        bpm:
          type: integer
          nullable: true
          description: Detected BPM rounded to nearest integer. Null if not embedded in file tags.
          example: 128
        bpm_label:
          type: string
          enum: [slow, medium, fast, very fast]
          description: Human-readable tempo category derived from BPM.
          example: "fast"
        key:
          type: string
          nullable: true
          description: Musical key (e.g. "C major", "A minor"). Null if not embedded.
          example: "C major"
        genre:
          type: array
          items:
            type: string
          description: Genre tags extracted from file metadata.
          example: ["Electronic", "Synthwave"]
        duration:
          type: integer
          nullable: true
          description: Track duration in seconds.
          example: 214
        format:
          type: string
          nullable: true
          description: Audio container format.
          example: "mp3"
        artist:
          type: string
          nullable: true
          description: Artist name from file tags.
          example: "My Band"
        title:
          type: string
          nullable: true
          description: Track title from file tags.
          example: "My Song"
        suggested_tags:
          type: string
          description: Ready-to-use Suno style tags string built from all detected attributes.
          example: "Electronic, Synthwave, 128 bpm, C major, fast"
        suggested_prompt:
          type: string
          description: Ready-to-use Suno description built from detected attributes.
          example: "Electronic, Synthwave style, 128 BPM, key of C major"
    GenerateFromAudioForm:
      type: object
      description: Multipart form fields for generate_from_audio.
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Your audio file (mp3, wav, flac, ogg, m4a …). Max 50 MB.
        title:
          type: string
          description: Song title. Falls back to the title embedded in the file, then "Generated Song".
        prompt:
          type: string
          description: Custom lyrics with section markers. Omit to generate without explicit lyrics.
        extra_tags:
          type: string
          description: Additional style tags appended to the auto-detected ones.
          example: "dark, cinematic"
        tags_override:
          type: string
          description: Fully replace the auto-detected tags with this value.
          example: "jazz, piano, slow"
        negative_tags:
          type: string
          description: Styles to exclude.
        model:
          type: string
          description: Suno model ID.
        wait_audio:
          type: string
          enum: ["true", "false"]
          default: "false"
          description: Pass "true" to block until audio URLs are ready.
        make_instrumental:
          type: string
          enum: ["true", "false"]
          default: "false"
          description: Pass "true" to generate without vocals.
        vocal_gender:
          type: string
          enum: [male, female]
        weirdness:
          type: string
          description: Integer 0–100 as a string.
        style_influence:
          type: string
          description: Integer 0–100 as a string.
        persona_id:
          type: string
          description: Voice-cloning persona UUID.
    GenerateFromAudioJson:
      type: object
      description: JSON body for generate_from_audio when providing a remote audio URL.
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Public URL of your audio file to analyze and use as style reference.
          example: "https://example.com/my-song.mp3"
        title:
          type: string
          description: Song title override.
        prompt:
          type: string
          description: Custom lyrics with section markers.
        extra_tags:
          type: string
          description: Additional tags appended to auto-detected ones.
          example: "dark, cinematic"
        tags_override:
          type: string
          description: Fully replace auto-detected tags.
          example: "jazz, piano, slow"
        negative_tags:
          type: string
          description: Styles to exclude.
        model:
          type: string
          description: Suno model ID.
        wait_audio:
          type: boolean
          default: false
        make_instrumental:
          type: boolean
          default: false
        vocal_gender:
          type: string
          enum: [male, female]
        weirdness:
          type: integer
          minimum: 0
          maximum: 100
        style_influence:
          type: integer
          minimum: 0
          maximum: 100
        persona_id:
          type: string
          format: uuid
    GenerateFromAudioResponse:
      type: object
      properties:
        analysis:
          $ref: '#/components/schemas/AudioAnalysis'
        clips:
          type: array
          items:
            $ref: '#/components/schemas/AudioInfo'
...
    AudioInfo:
      type: object
      description: Core audio clip object returned by all generation and retrieval endpoints.
      properties:
        id:
          type: string
          format: uuid
          description: Unique clip identifier.
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        title:
          type: string
          description: Song title.
          example: "Neon City Dreams"
        image_url:
          type: string
          format: uri
          description: Cover art image URL.
          example: "https://cdn2.suno.ai/image_a1b2c3d4.jpeg"
        lyric:
          type: string
          description: Full formatted lyrics with section markers.
        audio_url:
          type: string
          format: uri
          description: MP3 CDN URL. Available when `status` is `complete` or `streaming`.
          example: "https://cdn1.suno.ai/a1b2c3d4.mp3"
        video_url:
          type: string
          format: uri
          description: MP4 URL with waveform visualization. Available when `status` is `complete`.
          example: "https://cdn1.suno.ai/a1b2c3d4.mp4"
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
          example: "2026-04-17T12:00:00.000Z"
        model_name:
          type: string
          description: Model used for generation.
          example: "chirp-crow"
        gpt_description_prompt:
          type: string
          description: Internal GPT prompt used in auto-mode generation.
        prompt:
          type: string
          description: Original lyrics or description prompt submitted to Suno.
        status:
          type: string
          enum: [submitted, queued, streaming, complete, error]
          description: Current clip processing status.
          example: "complete"
        type:
          type: string
          description: Clip type identifier (e.g. `gen`, `stem`, `concat`).
        tags:
          type: string
          description: Style tags applied during generation.
          example: "synthwave, electronic, 80s"
        negative_tags:
          type: string
          description: Style tags that were excluded.
          example: "heavy metal, distortion"
        duration:
          type: string
          description: Clip duration in seconds (as a string).
          example: "185.4"
        error_message:
          type: string
          description: Human-readable error description. Only present when `status` is `error`.
        aligned_lyrics:
          type: array
          description: Word-level timing data. Populated by `/api/get_aligned_lyrics`.
          items:
            $ref: '#/components/schemas/AlignedWord'
    AlignedWord:
      type: object
      description: A single word with its start/end timestamps from lyric alignment.
      properties:
        word:
          type: string
          description: The lyric word.
          example: "City"
        start_s:
          type: number
          format: float
          description: Word start time in seconds.
          example: 12.34
        end_s:
          type: number
          format: float
          description: Word end time in seconds.
          example: 12.78
        success:
          type: boolean
          description: Whether alignment succeeded for this word.
        p_align:
          type: number
          format: float
          description: Alignment confidence probability (0–1).
          example: 0.97
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "Missing required parameter: prompt"

  # ── Reusable Responses ───────────────────────
  responses:
    BadRequest:
      description: Missing or invalid request parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Missing required parameter: prompt"
    InsufficientCredits:
      description: Insufficient Suno account credits
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Insufficient credits. Please upgrade your Suno plan."
    UnprocessableEntity:
      description: Invalid payload — usually a stale or invalid SUNO_COOKIE
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid session. Please refresh your SUNO_COOKIE."
    TooManyRequests:
      description: Rate limited by Suno API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Rate limited by Suno. Please try again later."
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "An unexpected error occurred."